app/requirements.txt
aiofiles==0.7.0
aioredis==2.0.0
aiosmtplib==1.1.6
alembic==1.6.5
anyio==3.3.0
asgiref==3.3.4
async-timeout==4.0.1
bcrypt==3.2.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.6.1
fastapi==0.70.0
fastapi-mail==1.0.2
google-auth==2.0.1
greenlet==1.1.0
h11==0.12.0
httpcore==0.13.6
httptools==0.2.0
httpx==0.19.0
idna==3.2
Jinja2==3.0.3
Mako==1.1.4
MarkupSafe==2.0.1
packaging==21.3
passlib==1.7.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.18
starlette==0.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
watchgod==0.7
websockets==9.1
psycopg2==2.9.3
